ajout InformationsController

+++ src/main/java/com/jseb23/NewPharmacie/Controller/PatientController.java
@@ -2,6 +2,7 @@ package com.jseb23.NewPharmacie.Controller;
 
 import com.jseb23.NewPharmacie.Model.Patient;
 import com.jseb23.NewPharmacie.Service.PatientService;
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -10,40 +11,63 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/patient")
+@RequestMapping("/patient")
+@AllArgsConstructor
 public class PatientController {
 
 
     private final PatientService patientService;
 
-    @Autowired
-    public PatientController(PatientService patientService) {
-        this.patientService = patientService;
-    }
-
-
-    @GetMapping("/all") // Récupère toutes les entités de type T
-    public List<Patient> getAllPatients() {
-        return patientService.getAll();
-    }
-
-    @GetMapping("/{id}") // Récupère une entité de type T par son identifiant
+    @GetMapping("/{id}")
     public ResponseEntity<Patient> getPatientById(@PathVariable Long id) {
-        Optional<Patient> patient = patientService.getById(id);
-        return patient.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        Optional<Patient> patient = patientService.findById(id);
+
+        return patient.map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/save") // Sauvegarde une entité de type T
-    public ResponseEntity<Patient> savePatient(@RequestBody Patient patient) {
-        Patient savedPatient = patientService.save(patient);
-        return ResponseEntity.ok(savedPatient);
+    @GetMapping("/all")
+    public ResponseEntity<List<Patient>> getAllPatients() {
+        List<Patient> patients = patientService.findAll();
+
+        if (!patients.isEmpty()) {
+            return ResponseEntity.ok(patients);
+        } else {
+            return ResponseEntity.noContent().build();
+        }
     }
 
+    @PostMapping("/create")
+    public ResponseEntity<Patient> createPatient(@RequestBody Patient patient) {
+        Patient createdPatient = patientService.save(patient);
+
+        return ResponseEntity.ok(createdPatient);
+    }
+
+    @PutMapping("/update/{id}")
+    public ResponseEntity<Patient> updatePatient(@PathVariable Long id, @RequestBody Patient patient) {
+        Optional<Patient> existingPatient = patientService.findById(id);
+
+        if (existingPatient.isPresent()) {
+            patient.setIdPatient(id);
+            Patient updatedPatient = patientService.save(patient);
+            return ResponseEntity.ok(updatedPatient);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deletePatient(@PathVariable Long id) {
-        patientService.deleteById(id);
-        return ResponseEntity.noContent().build();
+        Optional<Patient> existingPatient = patientService.findById(id);
+
+        if (existingPatient.isPresent()) {
+            patientService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
+
 
 }
